+++ CHANGELOG.md
@@ -117,6 +117,7 @@ the APIs in these libraries are stable, and are ready for production use.
 
 - [Essential Contacts](/google/cloud/essentialcontacts/README.md)
 - [Timeseries Insights](/google/cloud/timeseriesinsights/README.md)
+- [Verter AI](/google/cloud/aiplatform/README.md)
 
 ### Cloud Debugger